+++ src/engine.c
@@ -9,6 +9,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <math.h>
+#include "pmath.h"
 
 double now_s(void) { return plat_now(); }